Extend with additional GPX export
