+++ src/amazfit_exporter_config.py
@@ -1,0 +1,6 @@
+#!/usr/bin/python3
+
+activities = []
+trackpoint_data = []
+heart_rate_data = []
+x = 1
